requirements.txt
tabulate==0.8.2
knack==0.5.1
pandas==0.24.1
toolz==0.9.0
fire==0.1.3
azure-common==1.1.18
azure-cli-core==2.0.57
azureml-sdk==1.0.10
